内部通信 filter

+++ m-provider-user/src/main/java/com/radlly/model/User.java
@@ -11,7 +11,7 @@ public class User implements Serializable{
      */
     private static final long serialVersionUID = 5231134212346077681L;
 
-    private Long uuid;
+    private Integer uuid;
 
     @ApiModelProperty(value="用户名")  
     private String username;
@@ -32,11 +32,11 @@ public class User implements Serializable{
     private Byte status;
 
 
-	public Long getUuid() {
+	public Integer getUuid() {
 		return uuid;
 	}
 
-	public void setUuid(Long uuid) {
+	public void setUuid(Integer uuid) {
 		this.uuid = uuid;
 	}
